Refactor to extract lpsreach statistics to separate file.
